refactor: View의 책임을 입력과 출력으로 분리
