+++ src/test/java/calculator/CalculatorTest.java
@@ -1,12 +1,11 @@
 package calculator;
 
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import calculator.model.Calculator;
 import calculator.model.Parser;
-import calculator.model.Validator;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class CalculatorTest {
